requirements.txt
pandas
sqlalchemy
clickhouse-connect
python-dotenv 
